# 8360037: Refactor ImageReader in preparation for Valhalla support