--- a/src/java.base/share/classes/jdk/internal/jrtfs/ExplodedImage.java
+++ b/src/java.base/share/classes/jdk/internal/jrtfs/ExplodedImage.java
@@ -119,9 +119,9 @@ class ExplodedImage extends SystemImage {
         }
 
         @Override
-        public List<Node> getChildren() {
+        public Stream<String> getChildNames() {
             if (!isDirectory())
-                throw new IllegalArgumentException("not a directory: " + getNameString());
+                throw new IllegalArgumentException("not a directory: " + getName());
             if (children == null) {
                 List<Node> list = new ArrayList<>();
                 try (DirectoryStream<Path> stream = Files.newDirectoryStream(path)) {
@@ -138,7 +138,7 @@ class ExplodedImage extends SystemImage {
                 }
                 children = list;
             }
-            return children;
+            return children.stream().map(Node::getName);
         }
 
         @Override