… src/java.base/share/classes/jdk/internal/jrtfs/JrtFileSystem.java
… src/java.base/share/classes/jdk/internal/module/SystemModuleFinders.java
… test/jdk/jdk/internal/jimage/ImageReaderTest.java
… test/jdk/jdk/internal/jimage/JImageReadTest.java
… test/jdk/tools/jimage/ImageReaderDuplicateChildNodesTest.java
… test/micro/org/openjdk/bench/jdk/internal/jrtfs/ImageReaderBenchmark.java