@david-beaumont 2025-07-02
Latest changes uploaded. I will be unable to look at this for about a week now, but there are still some "TODO" comments in the ImageReader code which I'd like people's opinions on (esp. the possibly non-thread safe call to close() from SharedImageReader (not locked on the instance).