@david-beaumont 2025-07-16
Sadly it was much easier to force push here then jump through a lot of hoops (dependent PRs are a pain).
This is now the same code minus the ImageReaderBenchmark PR (though there are still 2 lines of changes in the .benchmark to update to the changed Node API). Other than that it's identical to before the push.