@AlanBateman 2025-07-31
I've done a number of passes over this and overall I think it looks quite good, and gives us a better encapsulated jimage API for jrtfs, the system module finder, and the jrt protocol handler to use. Thanks for taking all the feedback during all the iterations of this.